+++ tests/testthat/test-utilities.R
@@ -20,60 +20,95 @@ test_that('prepositions make good "or" patterns', {
 })
 
 
-test_that("match_pattern() works as expected for string matches", {
+test_that("match_by_pattern() works as expected for string matches", {
   labels <- c("Production [of b in c]", "d [of Coal in f]", "g [of h in USA]")
   # Simple matching
-  expect_equal(match_pattern(labels,
+  expect_equal(match_by_pattern(labels,
                              regex_pattern = "Production"),
                c(TRUE, FALSE, FALSE))
-  expect_equal(match_pattern(labels,
+  expect_equal(match_by_pattern(labels,
                              regex_pattern = "Coal"),
                c(FALSE, TRUE, FALSE))
-  expect_equal(match_pattern(labels,
+  expect_equal(match_by_pattern(labels,
                              regex_pattern = "USA"),
                c(FALSE, FALSE, TRUE))
 
   # Check word positions
-  expect_equal(match_pattern(labels,
+  expect_equal(match_by_pattern(labels,
                              regex_pattern = "^Production"),
                c(TRUE, FALSE, FALSE))
   # This should fail, because Production is at the start of the first string,
   # not the end of a string.
-  expect_equal(match_pattern(labels,
+  expect_equal(match_by_pattern(labels,
                              regex_pattern = "Production$"),
                c(FALSE, FALSE, FALSE))
 })
 
 
-test_that("match_pattern() works for prefixes and suffixes", {
+test_that("match_by_pattern() works for prefixes and suffixes", {
   labels <- c("Production [of b in c]", "d [of Coal in f]", "g [of h in USA]")
+
+  # This should give an error,
+  # because we should have only "pref" or "suff", not both.
+  expect_error(match_by_pattern(labels,
+                             regex_pattern = "Production",
+                             pieces = c("pref", "to")),
+               'If pieces contains "pref" or "suff", its length')
+
   # This should work, because "Production" is in the prefix.
-  expect_equal(match_pattern(labels,
+  expect_equal(match_by_pattern(labels,
                              regex_pattern = "Production",
                              pieces = "pref"),
                c(TRUE, FALSE, FALSE))
 
   # This should fail, because "Production" is not in the suffix.
-  expect_equal(match_pattern(labels,
+  expect_equal(match_by_pattern(labels,
                              regex_pattern = "Production",
                              pieces = "suff"),
                c(FALSE, FALSE, FALSE))
 })
 
 
-test_that("match_pattern() works for nouns and prepositions", {
+test_that("match_by_pattern() works for nouns and prepositions", {
   labels <- c("Production [of b in c]", "d [of Coal in f]", "g [of h in USA]")
   # This should work, because "Production" is a noun.
-  expect_equal(match_pattern(labels,
+  expect_equal(match_by_pattern(labels,
                              regex_pattern = "Production",
                              pieces = "noun"),
                c(TRUE, FALSE, FALSE))
 
   # This won't work, because "Production" is a noun, not in a prepositional phrase.
-  expect_equal(match_pattern(labels,
+  expect_equal(match_by_pattern(labels,
                              regex_pattern = "Production",
                              pieces = "in"),
                c(FALSE, FALSE, FALSE))
 
+  # Try a preposition
+  expect_equal(match_by_pattern(labels,
+                             regex_pattern = make_or_pattern(c("c", "f")),
+                             pieces = "in"),
+               c(TRUE, TRUE, FALSE))
 
+  # This should match only the USA one.
+  expect_equal(match_by_pattern(labels,
+                             regex_pattern = make_or_pattern(c("b", "Coal", "USA")),
+                             pieces = "in"),
+               c(FALSE, FALSE, TRUE))
+
+  # This should match all labels.
+  expect_equal(match_by_pattern(labels,
+                             regex_pattern = make_or_pattern(c("b", "Coal", "USA")),
+                             pieces = c("of", "in")),
+               c(TRUE, TRUE, TRUE))
+})
+
+
+test_that("match_by_pattern() works for degenerate case", {
+  labels <- c("Production [of b to GBR in c]", "d [of Coal in f]", "g [of h in USA]")
+  # Try a situation that will return non-square results.
+  expect_equal(match_by_pattern(labels,
+                             regex_pattern = make_or_pattern(c("b", "Coal", "GBR", "USA")),
+                             pieces = c("noun", "of", "in", "to"),
+                             prepositions = c("of", "to", "in")),
+               c(TRUE, TRUE, TRUE))
 })